Moved Chat classes to com.pengrad.telegrambot.model.chat
