+++ library/src/main/java/com/pengrad/telegrambot/model/MessageReactionCountUpdated.java
@@ -1,5 +1,6 @@
 package com.pengrad.telegrambot.model;
 
+import com.pengrad.telegrambot.model.chat.Chat;
 import com.pengrad.telegrambot.model.reaction.ReactionCount;
 
 import java.util.Arrays;
