+++ library/src/main/java/com/pengrad/telegrambot/model/chat/ChatLocation.java
@@ -1,4 +1,6 @@
-package com.pengrad.telegrambot.model;
+package com.pengrad.telegrambot.model.chat;
+
+import com.pengrad.telegrambot.model.Location;
 
 import java.io.Serializable;
 import java.util.Objects;
