+++ library/src/main/java/com/pengrad/telegrambot/model/chat/ChatMemberUpdated.java
@@ -1,4 +1,6 @@
-package com.pengrad.telegrambot.model;
+package com.pengrad.telegrambot.model.chat;
+
+import com.pengrad.telegrambot.model.User;
 
 import java.io.Serializable;
 import java.util.Objects;
