+++ library/src/main/java/com/pengrad/telegrambot/model/chat/boost/ChatBoost.java
@@ -1,6 +1,6 @@
-package com.pengrad.telegrambot.model.chatboost;
+package com.pengrad.telegrambot.model.chat.boost;
 
-import com.pengrad.telegrambot.model.chatboost.source.ChatBoostSource;
+import com.pengrad.telegrambot.model.chat.boost.source.ChatBoostSource;
 
 import java.util.Objects;
