+++ library/src/main/java/com/pengrad/telegrambot/model/message/origin/MessageOriginChat.java
@@ -1,6 +1,6 @@
 package com.pengrad.telegrambot.model.message.origin;
 
-import com.pengrad.telegrambot.model.Chat;
+import com.pengrad.telegrambot.model.chat.Chat;
 
 import java.util.Objects;
