+++ library/src/main/java/com/pengrad/telegrambot/model/request/KeyboardButtonRequestChat.java
@@ -2,7 +2,7 @@ package com.pengrad.telegrambot.model.request;
 
 import java.io.Serializable;
 
-import com.pengrad.telegrambot.model.ChatAdministratorRights;
+import com.pengrad.telegrambot.model.chat.ChatAdministratorRights;
 
 public class KeyboardButtonRequestChat implements Serializable {
     private final static long serialVersionUID = 0L;
