+++ library/src/main/java/com/pengrad/telegrambot/response/ChatInviteLinkResponse.java
@@ -1,6 +1,6 @@
 package com.pengrad.telegrambot.response;
 
-import com.pengrad.telegrambot.model.ChatInviteLink;
+import com.pengrad.telegrambot.model.chat.ChatInviteLink;
 
 /**
  * Stas Parshin
