+++ library/src/main/java/com/pengrad/telegrambot/response/GetChatAdministratorsResponse.java
@@ -1,6 +1,6 @@
 package com.pengrad.telegrambot.response;
 
-import com.pengrad.telegrambot.model.ChatMember;
+import com.pengrad.telegrambot.model.chat.ChatMember;
 
 import java.util.List;
